🚚rename: 슬라이스 폴더 이름을 케밥 케이스로 변경
