+++ src/widgets/forum-nav/ui/Category.tsx
