+++ src/widgets/hot-posts/ui/HotPostsCarousel.tsx
