+++ src/widgets/index.ts
@@ -1,11 +1,11 @@
-export { default as Category } from './forumNav/ui/Category';
-export { default as ForumNav } from './forumNav/ui/ForumNav';
-export { default as ForumPostList } from './forumPostList/ui/ForumPostList';
-export { default as HotPostsSlider } from './hotPosts/ui/HotPostsCarousel';
+export { default as Category } from './forum-nav/ui/Category';
+export { default as ForumNav } from './forum-nav/ui/ForumNav';
+export { default as ForumPostList } from './forum-post-list/ui/ForumPostList';
+export { default as HotPostsSlider } from './hot-posts/ui/HotPostsCarousel';
 export { default as GoBackButton } from '../shared/ui/GoBackButton';
-export { default as useForumPosts } from './forumPostList/model/useForumPosts';
-export { FORUM_CATEGORIES } from './forumNav/types/forumCategory';
-export { default as PostSearchBar } from './postSearchBar/ui/PostSearchBar';
+export { default as useForumPosts } from './forum-post-list/model/useForumPosts';
+export { FORUM_CATEGORIES } from './forum-nav/types/forumCategory';
+export { default as PostSearchBar } from './post-search-bar/ui/PostSearchBar';
 export { default as WriteForm } from './write-post/ui/WriteForm';
-export { default as PostDetail } from './postDetail/ui/PostDetail';
-export type { Post } from './postDetail/types/post';
+export { default as PostDetail } from './post-detail/ui/PostDetail';
+export type { Post } from './post-detail/types/post';
