+++ src/widgets/post-detail/api/fetchComments.ts
