+++ src/widgets/post-detail/model/usePost.ts
