+++ src/widgets/post-detail/types/comment.ts
