+++ src/widgets/post-detail/types/post.ts
