+++ src/widgets/post-detail/ui/Comment.tsx
