+++ src/widgets/post-detail/ui/CommentInput.tsx
