+++ src/widgets/post-detail/ui/MainContent.tsx
