+++ src/widgets/post-detail/ui/PostDetail.tsx
