+++ src/widgets/post-search-bar/ui/PostSearchBar.tsx
